feat: update user/room delete

+++ services/ws.js
@@ -1,30 +1,37 @@
-const { getRoomById } = require("./room");
-const { getUser, getRoomUsers, validateToken } = require("./user");
-const { getAllRoomVotes, getUserVoteInRoom, createVote, updateVote, getRoomMedian, deleteAllRoomVotes } = require("./vote");
+const { getRoomById, deleteRoom } = require("./room");
+const { getUser, getRoomUsers, deleteUser } = require("./user");
+const {
+	getAllRoomVotes,
+	getUserVoteInRoom,
+	createVote,
+	updateVote,
+	getRoomMedian,
+	deleteAllRoomVotes,
+} = require("./vote");
 
 const webSocketHandler = (io, socket) => {
-	socket.on('join', async ({roomId}) => {
-		const room = await getRoomById(roomId)
+	socket.on("join", async ({ roomId }) => {
+		const room = await getRoomById(roomId);
 		const user = await getUser(socket.decoded.userId);
 		console.log(`${user.name} login to room ${room.name}`);
 		socket.join(roomId);
 
 		// refresh users count to all users
-		io.to(roomId).emit('updateUserCount', {
+		io.to(roomId).emit("updateUserCount", {
 			data: {
-				users: await getRoomUsers(roomId)
-			}
-		})
+				users: await getRoomUsers(roomId),
+			},
+		});
 
 		// update votes
-		io.to(roomId).emit('updateVotes', {
+		io.to(roomId).emit("updateVotes", {
 			data: {
-				votes: await getAllRoomVotes(roomId)
-			}
+				votes: await getAllRoomVotes(roomId),
+			},
 		});
 	});
 
-	socket.on('vote', async ({roomId, userId, voteValue}) => {
+	socket.on("vote", async ({ roomId, userId, voteValue }) => {
 		const existedVote = await getUserVoteInRoom(userId, roomId);
 		if (!existedVote) {
 			await createVote(voteValue, roomId, userId);
@@ -33,27 +40,47 @@ const webSocketHandler = (io, socket) => {
 		}
 
 		// update votes
-		io.to(roomId).emit('updateVotes', {
+		io.to(roomId).emit("updateVotes", {
 			data: {
-				votes: await getAllRoomVotes(roomId)
-			}
+				votes: await getAllRoomVotes(roomId),
+			},
 		});
 	});
 
-	socket.on('showResults', async ({roomId}) => {
+	socket.on("showResults", async ({ roomId }) => {
 		// send event to all users
-		io.to(roomId).emit('showResults', {
+		io.to(roomId).emit("showResults", {
 			data: {
 				median: await getRoomMedian(roomId),
-			}
+			},
 		});
 	});
 
-	socket.on('reset', async ({roomId}) => {
-		await deleteAllRoomVotes(roomId)
+	socket.on("reset", async ({ roomId }) => {
+		await deleteAllRoomVotes(roomId);
 
-		io.to(roomId).emit('reset')
-	})
+		io.to(roomId).emit("reset");
+	});
+
+	socket.on("leftRoom", async ({ roomId, userId }) => {
+		await deleteUser(userId);
+
+		const usersLeft = await getRoomUsers(roomId);
+
+		// if there is no participants, delete room and all its votes
+		if (usersLeft.length === 0) {
+			deleteAllRoomVotes(roomId);
+			deleteRoom(roomId);
+			return;
+		}
+
+		// refresh users count to all users
+		io.to(roomId).emit("updateUserCount", {
+			data: {
+				users: await getRoomUsers(roomId),
+			},
+		});
+	});
 
 	// socket.on("join", ({ room, username }) => {
 	// 	socket.join(room);
@@ -83,17 +110,4 @@ const webSocketHandler = (io, socket) => {
 	});
 };
 
-const webSocketMiddleWare = (socket, next) => {
-	if (socket.handshake.query && socket.handshake.query.jwt) {
-		const decoded = validateToken(socket.handshake.query.jwt);
-		console.log('decoded', decoded)
-		if (!decoded) return next(new Error('Auth error from WS'));
-
-		socket.decoded = decoded;
-		next();
-	} else {
-		next(new Error('Auth error from WS'));
-	}
-}
-
-module.exports = { webSocketHandler, webSocketMiddleWare };
+module.exports = { webSocketHandler };
